Improved test framework with almost no duplicated server code
