+++ server/controllers/brogue-controller.js
@@ -13,6 +13,7 @@ var brogueComms = require('../brogue/brogue-comms');
 var gameRecordSchema = require('../database/game-record-model');
 var brogueConstants = require('../brogue/brogue-constants.js');
 
+var mongoose = require('mongoose');
 var gameRecord = mongoose.model('GameRecord', gameRecordSchema);
 
 // Controller for handling I/O with brogue process and client.  Note that unlike other controllers this one deals in binary data. Any incoming or outgoing binary data from this server should only come from this controller.
